# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: manga
    description: Manga search and chapter aggregation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /manga/aggregate:
    get:
      operationId: searchMangaAggregate
      tags: [manga]
      summary: Search manga using the primary chapter aggregator
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: Comick-backed search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaAggregate"
        "400":
          description: Invalid query
  /manga/comick/chapters:
    get:
      operationId: listComickChapters
      tags: [manga]
      summary: List chapters from Comick
      parameters:
        - name: slug
          in: query
          required: true
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
            default: en
        - name: limit
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterList"
        "502":
          description: Comick unavailable
  /manga/chapters:
    get:
      operationId: listMangaChapters
      tags: [manga]
      summary: List chapters using Comick first and MangaDex as fallback
      parameters:
        - name: comickSlug
          in: query
          schema:
            type: string
        - name: mangadexId
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
            default: en
        - name: limit
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Chapters from the first available source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterList"
        "400":
          description: A source identifier is required
        "502":
          description: Both chapter sources are unavailable
  /manga/mangadex/chapters:
    get:
      operationId: listMangaDexChapters
      tags: [manga]
      summary: List fallback chapters from MangaDex
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
            default: en
        - name: limit
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: MangaDex chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterList"
        "400":
          description: MangaDex id is required
        "502":
          description: MangaDex unavailable
  /manga/mangaupdates/search:
    get:
      operationId: searchMangaUpdates
      tags: [manga]
      summary: Search MangaUpdates metadata
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: MangaUpdates series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaUpdatesSearchResponse"
        "400":
          description: Invalid query
        "502":
          description: MangaUpdates unavailable
  /manga/mangaupdates/series:
    get:
      operationId: getMangaUpdatesSeries
      tags: [manga]
      summary: Get MangaUpdates series metadata
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: MangaUpdates series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaUpdatesSeries"
        "404":
          description: Series not found
        "502":
          description: MangaUpdates unavailable
  /manga/mangaupdates/releases:
    get:
      operationId: searchMangaUpdatesReleases
      tags: [manga]
      summary: Search MangaUpdates releases
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: MangaUpdates releases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaUpdatesReleaseResponse"
        "400":
          description: Invalid query
        "502":
          description: MangaUpdates unavailable
  /manga/mangaupdates/tracking:
    get:
      operationId: getMangaUpdatesTracking
      tags: [manga]
      summary: Get MangaUpdates tracking snapshot
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaUpdatesTrackingSnapshot"
        "404":
          description: Series not found
        "502":
          description: MangaUpdates unavailable
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions
      description: Returns all active MAL subscriptions, optionally filtered by guild
      parameters:
        - in: query
          name: guild_id
          schema:
            type: string
          description: Filter by Discord guild ID
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Subscribe to a MAL item
      description: Subscribes a Discord user to receive notifications for a MAL manga/anime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscriptions/{id}:
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Remove a subscription
      description: Removes a subscription by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscriptions/{id}/check:
    post:
      operationId: checkSubscription
      tags: [subscriptions]
      summary: Check for updates on a subscription
      description: >
        Fetches the latest data from MAL for this subscription, stores a snapshot
        (synopsis, score, status, chapters), and returns whether chapters changed.
        Keeps only the 2 most recent snapshots — oldest is deleted automatically.
        Only returns changed=true when the chapter count increases.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResult"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: MAL API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /errors:
    get:
      operationId: listErrorLogs
      tags: [errors]
      summary: List bot errors
      description: Returns recent errors, optionally filtered by Discord guild.
      parameters:
        - in: query
          name: guild_id
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Recent error history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorLog"
    post:
      operationId: createErrorLog
      tags: [errors]
      summary: Record a bot error
      description: Stores an error reported by the Discord bot in Neon for later inspection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErrorLogInput"
      responses:
        "201":
          description: Error recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorLog"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MangaRecord:
      type: object
      required: [source, id, title, alternativeTitles, description, coverUrl, score, genres, status, chapters, year, country, url]
      properties:
        source:
          type: string
          enum: [comick, mangaupdates, mangadex, jikan, anilist]
        id: { type: string }
        title: { type: string }
        alternativeTitles:
          type: array
          items: { type: string }
        description: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        score: { type: ["number", "null"] }
        genres:
          type: array
          items: { type: string }
        status: { type: ["string", "null"] }
        chapters: { type: ["number", "null"] }
        year: { type: ["number", "null"] }
        country: { type: ["string", "null"] }
        url: { type: string }

    SourceStatus:
      type: object
      required: [source, status, count]
      properties:
        source:
          type: string
          enum: [comick, mangaupdates, mangadex, jikan, anilist]
        status:
          type: string
          enum: [ok, empty, error]
        count: { type: number }
        error: { type: string }

    MangaAggregate:
      type: object
      required: [query, primarySource, results, sources]
      properties:
        query: { type: string }
        primarySource:
          type: string
          enum: [comick]
        results:
          type: array
          items:
            $ref: "#/components/schemas/MangaRecord"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceStatus"

    Chapter:
      type: object
      required: [id, chapter, volume, title, language, publishedAt, url, group, source]
      properties:
        id: { type: string }
        chapter: { type: string }
        volume: { type: ["string", "null"] }
        title: { type: ["string", "null"] }
        language: { type: ["string", "null"] }
        publishedAt: { type: ["string", "null"] }
        url: { type: ["string", "null"] }
        group: { type: ["string", "null"] }
        source:
          type: string
          enum: [comick, mangadex]

    ChapterList:
      type: object
      required: [source, mangaId, chapters]
      properties:
        source:
          type: string
          enum: [comick, mangadex]
        mangaId: { type: string }
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/Chapter"

    MangaUpdatesGroup:
      type: object
      required: [id, name, url]
      properties:
        id: { type: ["string", "null"] }
        name: { type: string }
        url: { type: ["string", "null"] }

    MangaUpdatesSeries:
      allOf:
        - $ref: "#/components/schemas/MangaRecord"
        - type: object
          required: [sourceRole, latestChapter, statusText, groups, lastUpdated]
          properties:
            sourceRole:
              type: string
              enum: [metadata-and-releases]
            latestChapter: { type: ["number", "null"] }
            statusText: { type: ["string", "null"] }
            groups:
              type: array
              items:
                $ref: "#/components/schemas/MangaUpdatesGroup"
            lastUpdated: { type: ["string", "null"] }

    MangaUpdatesSearchResponse:
      type: object
      required: [source, results]
      properties:
        source:
          type: string
          enum: [mangaupdates]
        results:
          type: array
          items:
            $ref: "#/components/schemas/MangaUpdatesSeries"

    MangaUpdatesRelease:
      type: object
      required: [id, title, volume, chapter, groups, releaseDate, addedAt, url]
      properties:
        id: { type: string }
        title: { type: string }
        volume: { type: ["string", "null"] }
        chapter: { type: string }
        groups:
          type: array
          items:
            $ref: "#/components/schemas/MangaUpdatesGroup"
        releaseDate: { type: ["string", "null"] }
        addedAt: { type: ["string", "null"] }
        url: { type: string }

    MangaUpdatesReleaseResponse:
      type: object
      required: [source, query, releases]
      properties:
        source:
          type: string
          enum: [mangaupdates]
        query: { type: string }
        releases:
          type: array
          items:
            $ref: "#/components/schemas/MangaUpdatesRelease"

    MangaUpdatesTrackingSnapshot:
      type: object
      required: [source, seriesId, title, latestChapter, status, statusText, lastUpdated, checkedAt, chaptersAreOfficialSource]
      properties:
        source:
          type: string
          enum: [mangaupdates]
        seriesId: { type: string }
        title: { type: string }
        latestChapter: { type: ["number", "null"] }
        status: { type: ["string", "null"] }
        statusText: { type: ["string", "null"] }
        lastUpdated: { type: ["string", "null"] }
        checkedAt: { type: string }
        chaptersAreOfficialSource:
          type: boolean
          enum: [false]


    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        error_id:
          type: number


    ErrorLog:
      type: object
      required:
        - id
        - error_code
        - message
        - created_at
      properties:
        id:
          type: number
        discord_guild_id:
          type: ["string", "null"]
        discord_user_id:
          type: ["string", "null"]
        command:
          type: ["string", "null"]
        route:
          type: ["string", "null"]
        error_code:
          type: string
        message:
          type: string
        http_status:
          type: ["number", "null"]
        context:
          type: ["object", "null"]
          additionalProperties: true
        created_at:
          type: string
          format: date-time


    ErrorLogInput:
      type: object
      required:
        - error_code
        - message
      properties:
        discord_guild_id:
          type: string
        discord_user_id:
          type: string
        command:
          type: string
        route:
          type: string
        error_code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 2000
        http_status:
          type: number
        context:
          type: object
          additionalProperties: true


    Subscription:
      type: object
      required:
        - id
        - discord_user_id
        - discord_channel_id
        - discord_guild_id
        - mal_item_id
        - mal_item_type
        - item_name
        - created_at
      properties:
        id:
          type: number
        discord_user_id:
          type: string
        discord_channel_id:
          type: string
        discord_guild_id:
          type: string
        mal_item_id:
          type: number
        mal_item_type:
          type: string
          enum: [manga, anime]
        item_name:
          type: string
        created_at:
          type: string
          format: date-time


    SubscriptionInput:
      type: object
      required:
        - discord_user_id
        - discord_channel_id
        - discord_guild_id
        - mal_item_id
        - mal_item_type
        - item_name
      properties:
        discord_user_id:
          type: string
        discord_channel_id:
          type: string
        discord_guild_id:
          type: string
        mal_item_id:
          type: number
        mal_item_type:
          type: string
          enum: [manga, anime]
        item_name:
          type: string


    CheckResult:
      type: object
      required:
        - changed
        - subscription_id
        - item_name
        - mal_item_id
        - mal_item_type
        - current
      properties:
        changed:
          type: boolean
          description: True only when chapter/episode count increased
        subscription_id:
          type: number
        item_name:
          type: string
        mal_item_id:
          type: number
        mal_item_type:
          type: string
        current:
          $ref: "#/components/schemas/SnapshotData"
        previous:
          oneOf:
            - $ref: "#/components/schemas/SnapshotData"
            - type: "null"


    SnapshotData:
      type: object
      required:
        - synopsis
        - score
        - status
        - chapters
        - checked_at
      properties:
        synopsis:
          type: ["string", "null"]
        score:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
        chapters:
          type: ["number", "null"]
        checked_at:
          type: string
          format: date-time